+++ src/main/java/hu/oe/nik/szfmv/automatedcar/AutomatedCar.java
@@ -1,9 +1,9 @@
-package hu.oe.nik.szfmv17t.automatedcar;
+package hu.oe.nik.szfmv.automatedcar;
 
 
-import hu.oe.nik.szfmv17t.automatedcar.bus.VirtualFunctionBus;
-import hu.oe.nik.szfmv17t.automatedcar.powertrainsystem.PowertrainSystem;
-import hu.oe.nik.szfmv17t.environment.WorldObject;
+import hu.oe.nik.szfmv.automatedcar.bus.VirtualFunctionBus;
+import hu.oe.nik.szfmv.automatedcar.powertrainsystem.PowertrainSystem;
+import hu.oe.nik.szfmv.environment.WorldObject;
 
 public class AutomatedCar extends WorldObject{
