+++ src/main/java/hu/oe/nik/szfmv/automatedcar/SystemComponent.java
@@ -1,6 +1,6 @@
-package hu.oe.nik.szfmv17t.automatedcar;
+package hu.oe.nik.szfmv.automatedcar;
 
-import hu.oe.nik.szfmv17t.automatedcar.bus.VirtualFunctionBus;
+import hu.oe.nik.szfmv.automatedcar.bus.VirtualFunctionBus;
 
 /**
  *
